libs/langgraph/pyproject.toml
[tool.poetry]
name = "langchain-nvidia-langgraph"
version = "1.0.4"
description = "NVIDIA-optimized LangGraph using a set of foundational agent performance primitives"
authors = []
readme = "README.md"
repository = "https://github.com/langchain-ai/langchain-nvidia"
license = "MIT"
packages = [
  { include = "langchain_nvidia_langgraph" },
]

[tool.poetry.urls]
"Source Code" = "https://github.com/langchain-ai/langchain-nvidia/tree/main/libs/langgraph"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
langchain-core = ">=1.2.5,<2.0.0"
langgraph = ">=1.0.0,<2.0.0"
nvidia-nat-app = { version = "^1.5.0", allow-prereleases = true }

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.3.0"
pytest-mock = "^3.10.0"
pytest-watcher = "^0.3.4"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.codespell]
optional = true

[tool.poetry.group.codespell.dependencies]
codespell = "^2.2.0"

[tool.poetry.group.test_integration]
optional = true

[tool.poetry.group.test_integration.dependencies]

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = "^0.1.5"

[tool.poetry.group.typing.dependencies]
mypy = "^0.991"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle
  "F",    # pyflakes
  "I",    # isort
  "T201", # print
]

[tool.mypy]
disallow_untyped_defs = "True"
exclude = ["notebooks", "examples", "example_data"]

[[tool.mypy.overrides]]
# conditional dependencies introduced by langsmith-sdk
module = ["numpy", "pytest"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# nat_app has no type stubs
module = [
  "nat_app",
  "nat_app.*",
]
ignore_missing_imports = true

[tool.coverage.run]
omit = ["tests/*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
# --strict-markers will raise errors on unknown marks.
# https://docs.pytest.org/en/7.1.x/how-to/mark.html#raising-errors-on-unknown-marks
#
# https://docs.pytest.org/en/7.1.x/reference/reference.html
# --strict-config       any warnings encountered while parsing the `pytest`
#                       section of the configuration file raise errors.
#
# https://github.com/tophat/syrupy
# --snapshot-warn-unused    Prints a warning on unused snapshots rather than fail the test suite.
addopts = "--strict-markers --strict-config --durations=5"
# Registering custom markers.
# https://docs.pytest.org/en/7.1.x/example/markers.html#registering-markers
markers = [
  "asyncio: mark tests as requiring asyncio",
  "integration: mark integration tests (end-to-end compile and invoke)",
]
asyncio_mode = "auto"
